Update tandy sound core to latest (unmerged) mame version. Update tandy sound selection to use the ncr8496 for normal usage and the sn76496 in PCjr mode.
Correct noise algorithm. Ignore writes to certain registers if highest bit is not set for ncr8496 mode. Don't reset the noise producing register on all writes, only when changing the type of noise (ncr8496). 
Fixes intro sound of KQ2, waves in KQ2 and sound on the title screen of sentinel.
Thanks for the help Lord_Nightmare. 

git-svn-id: https://svn.code.sf.net/p/dosbox/code-0/dosbox/trunk@4141 ed541006-0bf7-43e9-8c4d-6fc63c346d47